+++ stringutils.cpp
@@ -79,7 +79,7 @@ vector<string> explode( const string &delimiter, const string &str)
 
 string merge(const vector<string>& v, const string& delim) {
     string res;
-    for (int i = 0; i < v.size(); ++i) {
+    for (int i = 0; i < (int)v.size(); ++i) {
         if (i) {
             res += delim;
         }
